Added a ratings widget on the post. Also implemented a categories table to select a category. Updated for iOS11.3

+++ Places/RPlace.swift
@@ -18,8 +18,8 @@ class RPlace: Mappable {
     var defaultImage: String?
     var category: String?
     var images: [String]?
-    var location: [Float]?
-    var rating: Float?
+    var location: [Double]?
+    var rating: Double?
     
     required init(map: Map){
         self.defaultImage = "defaultPhoto"
